Storing changed content before calling hooks
So the hook programs can diff (or sha1sum in the sample file) both content files.

But I'm not very happy with the variable names (old_filename, "$old_filename.old").

Besides that, there is always a filename provided for the old content, even when the content didn't change. For that, the sha1sum fails on the non-existing (old) file.